feat: Add gen.yaml generation.mockServer configuration
Reference: https://linear.app/speakeasy/issue/SPE-4016/feature-add-mockserver-generation-configuration-flag
Reference: https://linear.app/speakeasy/issue/SPE-4011/feature-enable-mockserver-generation-for-customers

Introduces a new `mockServer` configuration object under the `generation` object that will be used to configure generation settings for the mock server that will be eventually bundled with test generation. While the mock server is under development, the generator will explicitly look for the `MockServer` configuration object to be present and `disabled` to be `false`. When we want to enable mock server generation for customers, we will update that logic to treat missing `MockServer` as "enabled" or if the object is present, respect the `disabled` flag (effectively making it opt-out).