feat: Add operations information to gen.lock for change reporting
This change adds a `gen.lock` (`LockFile`) type `Operations` field for API and SDK change reporting. This is intentionally a subset of operation data for prioritizing important information to present to API producers and consumers, while not exposing more of the generator internals. This information is intended to be self-contained to prevent the need for the generator or OAS-based tooling when generating API and SDK change reports, e.g. comparing the prior snapshot with a new one should contain everything necessary.

The operation representation metadata can be enhanced over time if additional SDK change reporting, such as new arguments, is desired.